Add test cases

+++ contracts/BlastLootBox.sol
@@ -27,21 +27,25 @@ contract BlastLootBox is
     using Counters for Counters.Counter;
 
     struct LootBox {
-        uint token0;
-        uint token1;
-        uint token2;
+        uint256 token0;
+        uint256 token1;
+        uint256 token2;
     }
 
     bytes32 public constant GAME_ROLE = keccak256("GAME_ROLE");
 
     Counters.Counter private _tokenIdCounter;
-    mapping(uint => LootBox) private lootboxDetails;
-    mapping(uint => uint8) private tokenTypes;
+    mapping(uint256 => LootBox) private lootboxDetails;
+    mapping(uint256 => uint8) private tokenTypes;
     IBlastEquipmentNFT public blastEquipmentNFT;
 
     /// @param name Name of the contract
     /// @param symbol Symbol of the contract
-    constructor(string memory name, string memory symbol, IBlastEquipmentNFT _blastEquipmentNFT) ERC721(name, symbol) {
+    constructor(
+        string memory name,
+        string memory symbol,
+        IBlastEquipmentNFT _blastEquipmentNFT
+    ) ERC721(name, symbol) {
         if (address(_blastEquipmentNFT) == address(0)) revert NoZeroAddress();
         _setupRole(DEFAULT_ADMIN_ROLE, _msgSender());
         _setupRole(GAME_ROLE, _msgSender());
@@ -51,14 +55,17 @@ contract BlastLootBox is
     /// @notice Creates a new token for `_to`. Its token ID will be automatically
     /// @dev The caller must have the `DEFAULT_ADMIN_ROLE`.
     /// _tokenType should be 1 or 2 (In case of 1, it's normal box. In case of 2, it's gw box)
-    function safeMint(address[] calldata _to, string[] calldata _uri, LootBox[] calldata _eqtIds, uint8 _tokenType)
-        external
-        onlyRole(DEFAULT_ADMIN_ROLE)
-    {
-        if (_to.length != _uri.length || _to.length != _eqtIds.length) revert InvalidParams();
+    function safeMint(
+        address[] calldata _to,
+        string[] calldata _uri,
+        LootBox[] calldata _eqtIds,
+        uint8 _tokenType
+    ) external onlyRole(DEFAULT_ADMIN_ROLE) {
+        if (_to.length != _uri.length || _to.length != _eqtIds.length)
+            revert InvalidParams();
         if (!(_tokenType == 1 || _tokenType == 2)) revert InvalidParams();
 
-        for (uint i = 0; i < _to.length; i++) {
+        for (uint256 i = 0; i < _to.length; i++) {
             uint256 tokenId = _tokenIdCounter.current();
             _tokenIdCounter.increment();
             lootboxDetails[tokenId] = _eqtIds[i];
@@ -68,21 +75,24 @@ contract BlastLootBox is
         }
     }
 
-    function open(uint _tokenId) external {
+    function open(uint256 _tokenId) external {
         if (!_exists(_tokenId)) revert NonExistToken();
         if (_msgSender() != ownerOf(_tokenId)) revert NotOwner();
 
         _open(_tokenId, _msgSender());
     }
 
-    function openTo(uint _tokenId, address _to) external onlyRole(GAME_ROLE) {
+    function openTo(uint256 _tokenId, address _to)
+        external
+        onlyRole(GAME_ROLE)
+    {
         if (!_exists(_tokenId)) revert NonExistToken();
         if (_to != ownerOf(_tokenId)) revert NotOwner();
 
         _open(_tokenId, _to);
     }
 
-    function _open(uint _tokenId, address _to) internal {
+    function _open(uint256 _tokenId, address _to) internal {
         LootBox memory _eqtIds = lootboxDetails[_tokenId];
 
         blastEquipmentNFT.transferFrom(address(this), _to, _eqtIds.token0);
@@ -100,7 +110,12 @@ contract BlastLootBox is
     /// @notice Get Token Type. (GWB or NB)
     /// @dev Returned value should be 1 or 2
     /// @param _tokenId Token ID.
-    function getTokenType(uint _tokenId) external view override returns (uint8) {
+    function getTokenType(uint256 _tokenId)
+        external
+        view
+        override
+        returns (uint8)
+    {
         return tokenTypes[_tokenId];
     }
